Рефакторинг приложения.
