+++ app/src/main/java/com/volokhinaleksey/androidmaps/room/PointDatabase.kt
@@ -3,6 +3,10 @@ package com.volokhinaleksey.androidmaps.room
 import androidx.room.Database
 import androidx.room.RoomDatabase
 
+/**
+ * Local Database object
+ */
+
 @Database(
     entities = [PointEntity::class],
     version = 1
